# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Internship Transcript System V2 API
  version: 0.2.0
  summary: Target contract for internship evaluation, correspondence, and transcript generation
  description: |
    Canonical target contract for V2. This specification is derived from observed legacy workflows
    but intentionally replaces legacy /api/v1 shapes, unauthenticated CRUD, mutable templates,
    and synchronous email/PDF behavior.

    Unresolved organization policies remain marked TBD in descriptions and acceptance criteria.
    No Production credential belongs in this file.
  contact:
    name: Internship Transcript System Owner
  license:
    name: Proprietary / TBD
    identifier: LicenseRef-Proprietary
servers:
  - url: https://{apiHost}/api/v2
    description: Production placeholder
    variables:
      apiHost:
        default: api.example.invalid
        description: Inject the approved Production API hostname
...
security:
  - cookieAuth: []
paths:
...
  /users:
    get:
      tags: [Auth]
      operationId: listUsers
      summary: List users within the caller's approved scope
      x-permissions: [users.read]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: search
          in: query
          schema: { type: string, maxLength: 100 }
        - name: role
          in: query
          schema:
            type: string
            enum: [systemAdmin, internshipStaff, coordinator, student, evaluator, auditor]
        - name: status
          in: query
          schema: { type: string, enum: [active, archived, suspended] }
        - name: schoolId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Scoped user page without OIDC subject identifiers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedUserPage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Auth]
      operationId: createUser
      summary: Create a scoped operational user or explicitly link a student record
      x-permissions: [users.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManagedUser'
      responses:
        '201':
          description: Created scoped user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedUser'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/summary:
    get:
      tags: [Auth]
      operationId: getUserSummary
      summary: Count users within the caller's management scope
      x-permissions: [users.manage]
      responses:
        '200':
          description: Scoped user totals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    patch:
      tags: [Auth]
      operationId: updateUser
      summary: Update a user within the caller's management scope
      x-permissions: [users.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateManagedUser'
      responses:
        '200':
          description: Updated scoped user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedUser'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Auth]
      operationId: archiveUser
      summary: Archive a user within the caller's management scope
      x-permissions: [users.manage]
      responses:
        '204':
          description: User archived
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: its_session
      description: Secure httpOnly same-site internal session cookie
    evaluatorBearer:
      type: http
      scheme: bearer
      bearerFormat: scoped-invitation-session
      description: Short-lived evaluator session bound to one assignment
...
  responses:
    Unauthenticated:
      description: Authentication is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: AUTHENTICATION_REQUIRED
            message: Authentication is required.
            requestId: req_example
...
    Forbidden:
      description: Authenticated caller lacks action or scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: PERMISSION_DENIED
            message: You do not have permission to perform this action.
            requestId: req_example
    NotFound:
      description: Resource does not exist or is concealed from caller
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESOURCE_NOT_FOUND
            message: Resource not found.
            requestId: req_example
    Conflict:
      description: Duplicate, stale version, idempotency conflict, or invalid state transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request or publish validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: One or more fields are invalid.
            details:
              - field: email
                reason: invalid_format
            requestId: req_example
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        requestId:
          type: string
    ErrorDetail:
      type: object
      required: [reason]
      properties:
        field:
          type: string
        reason:
          type: string
        message:
          type: string
...
    ManagedUser:
      type: object
      required: [id, email, displayName, status, roleAssignments]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
        status: { type: string, enum: [active, archived, suspended] }
        studentId: { type: string }
        roleAssignments:
          type: array
          items:
            type: object
            required: [role, tenant, schoolIds, programIds, active]
            properties:
              role:
                type: string
                enum: [systemAdmin, internshipStaff, coordinator, student, evaluator, auditor]
              tenant: { type: boolean }
              schoolIds:
                type: array
                items: { type: string }
              programIds:
                type: array
                items: { type: string }
              active: { type: boolean }
    ManagedUserPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ManagedUser'
        meta:
          type: object
          required: [page, pageSize, total, totalPages]
          properties:
            page: { type: integer, minimum: 1 }
            pageSize: { type: integer, minimum: 1 }
            total: { type: integer, minimum: 0 }
            totalPages: { type: integer, minimum: 0 }
    UserSummary:
      type: object
      required: [total, systemAdmin, internshipStaff, coordinator, student]
      properties:
        total: { type: integer, minimum: 0 }
        systemAdmin: { type: integer, minimum: 0 }
        internshipStaff: { type: integer, minimum: 0 }
        coordinator: { type: integer, minimum: 0 }
        student: { type: integer, minimum: 0 }
    CreateManagedUser:
      type: object
      required: [email, displayName, role]
      properties:
        email: { type: string, format: email }
        displayName: { type: string, minLength: 1, maxLength: 200 }
        role:
          type: string
          enum: [systemAdmin, internshipStaff, coordinator, student, evaluator, auditor]
        status: { type: string, enum: [active, archived, suspended] }
        studentId: { type: string, maxLength: 50 }
        schoolIds:
          type: array
          items: { type: string }
        programIds:
          type: array
          items: { type: string }
    UpdateManagedUser:
      type: object
      minProperties: 1
      properties:
        email: { type: string, format: email }
        displayName: { type: string, minLength: 1, maxLength: 200 }
        role:
          type: string
          enum: [systemAdmin, internshipStaff, coordinator, student, evaluator, auditor]
        status: { type: string, enum: [active, archived, suspended] }
        studentId: { type: string, maxLength: 50 }
        schoolIds:
          type: array
          items: { type: string }
        programIds:
          type: array
          items: { type: string }
...